[ENHANCEMENT] Deduplicate Recurring and NonRecurring management ViewModels (issue #11)
