+++ app/src/main/java/fr/laforge/benoist/financialmanager/presentation/ui/transaction/recurring/RecurringExpensesScreen.kt
@@ -22,7 +22,7 @@ fun RecurringExpensesScreen(
     vm: RecurringManagementViewModel = koinViewModel(),
 ) {
     // Collect Data
-    val transactions by vm.recurringExpensesItems.collectAsState()
+    val transactions by vm.expenseItems.collectAsState()
     val totalMonthly by vm.totalExpensesMonthly.collectAsState()
     val query by vm.query.collectAsState()
